separate showed and hidden addresses
