+++ lib/screen/wallet_connect/wc_connect_page.dart
@@ -13,7 +13,6 @@ import 'package:autonomy_flutter/model/connection_request_args.dart';
 import 'package:autonomy_flutter/screen/app_router.dart';
 import 'package:autonomy_flutter/screen/bloc/accounts/accounts_bloc.dart';
 import 'package:autonomy_flutter/screen/connection/persona_connections_page.dart';
-import 'package:autonomy_flutter/screen/wallet_connect/v2/wc2_permission_page.dart';
 import 'package:autonomy_flutter/service/configuration_service.dart';
 import 'package:autonomy_flutter/service/ethereum_service.dart';
 import 'package:autonomy_flutter/service/metric_client_service.dart';
@@ -29,11 +28,11 @@ import 'package:autonomy_flutter/util/style.dart';
 import 'package:autonomy_flutter/util/ui_helper.dart';
 import 'package:autonomy_flutter/util/wallet_storage_ext.dart';
 import 'package:autonomy_flutter/view/back_appbar.dart';
+import 'package:autonomy_flutter/view/list_address_account.dart';
 import 'package:autonomy_flutter/view/primary_button.dart';
 import 'package:autonomy_flutter/view/responsive.dart';
 import 'package:autonomy_theme/autonomy_theme.dart';
 import 'package:cached_network_image/cached_network_image.dart';
-import 'package:collection/collection.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -628,40 +627,22 @@ class _WCConnectPageState extends State<WCConnectPage>
           ),
         ),
         const SizedBox(height: 16.0),
-        Column(
-          children: [
-            ...accounts
-                .map((account) => [
-                      PersonalConnectItem(
-                        categorizedAccount: account,
-                        ethSelectedAddress: ethSelectedAddress,
-                        tezSelectedAddress: tezSelectedAddress,
-                        isExpand: true,
-                        onSelectEth: (value) async {
-                          int index =
-                              await account.persona?.getAddressIndex(value) ??
-                                  0;
-                          setState(() {
-                            ethSelectedAddress = value;
-                            selectedPersona =
-                                WalletIndex(account.persona!.wallet(), index);
-                          });
-                        },
-                        onSelectTez: (value) async {
-                          int index =
-                              await account.persona?.getAddressIndex(value) ??
-                                  0;
-                          setState(() {
-                            tezSelectedAddress = value;
-                            selectedPersona =
-                                WalletIndex(account.persona!.wallet(), index);
-                          });
-                        },
-                      ),
-                    ])
-                .flattened
-                .toList(),
-          ],
+        ListAccountConnect(
+          accounts: accounts,
+          onSelectEth: (value) {
+            int index = value.walletAddress?.index ?? 0;
+            setState(() {
+              ethSelectedAddress = value.accountNumber;
+              selectedPersona = WalletIndex(value.persona!.wallet(), index);
+            });
+          },
+          onSelectTez: (value) {
+            int index = value.walletAddress?.index ?? 0;
+            setState(() {
+              tezSelectedAddress = value.accountNumber;
+              selectedPersona = WalletIndex(value.persona!.wallet(), index);
+            });
+          },
         ),
       ],
     );
